refactor: 이상한 곳에 가잇는 파일 정리
