+++ apps/web/src/app/[locale]/shop/_auction/index.tsx
@@ -1,9 +1,9 @@
 import { type ReactNode } from 'react';
 import { css } from '_panda/css';
 
-import SellListSection from '../_sellList';
 import type { TabType } from '../type';
 
+import SellListSection from './SellSection/SellListSection';
 import { Background } from './Background';
 import { DefaultTabRight } from './DefaultTabRight';
 import HistoryTable from './HistoryTable';
